Big refactor of files through the system to organize it
