+++ library/llms/llm.py
@@ -1,6 +1,6 @@
 from langchain.callbacks.streaming_stdout import StreamingStdOutCallbackHandler
-from library.promptmanager import PromptManager
-from library.weaviate import VDB
+from library.llms.promptmanager import PromptManager
+from library.data.local.weaviate import VDB
 from langchain.chains.llm import LLMChain
 from langchain_community.llms import GPT4All
 from langchain import PromptTemplate
