+++ library/models/person.py
@@ -1,17 +1,19 @@
 import hashlib
-class Person:
-    def __init__(self, name, email):
-        self.name = name
-        self.email = email
 
-    def identifier(self):
+from pydantic import BaseModel
+class Person(BaseModel):
+    name: str
+    email: str
+
+    @property
+    def identifier(self) -> str:
         sha256 = hashlib.sha256()
         sha256.update(self.email.encode('utf-8'))
         return sha256.hexdigest()
 
     def to_dict(self):
         return {
-            "id": self.identifier(),
+            "id": self.identifier,
             "name": self.name,
             "email": self.email
         }
