+++ library/models/weaviate_schemas.py
