+++ tests/integration/library/test_importance_service.py
@@ -6,15 +6,15 @@ from typing import Optional
 import uuid
 import pytest
 import requests
-from library import weaviate as w
-from library.api_models import EmailMessage
-from library.employee import Employee
-import library.handlers as h
+from library.data.local import weaviate as w
+from library.models.api_models import EmailMessage
+from library.models.employee import Employee
+import library.managers.handlers as h
 from requests.exceptions import ConnectionError
 
-from library.importance import ImportanceService
+from library.managers.importance import ImportanceService
 from library.models.message import Message
-from library.weaviate_schemas import WeaviateSchemas
+from library.models.weaviate_schemas import WeaviateSchemas
 from tests.integration.library.integration_test_base import IntegrationTestBase
 
 class TestImportanceService(IntegrationTestBase):
