+++ tests/integration/library/test_slack_weaviate.py
@@ -2,11 +2,11 @@ import json
 import os
 import pytest
 import requests
-from library import weaviate as w
-import library.handlers as h
+from library.data.local import weaviate as w
+import library.managers.handlers as h
 from requests.exceptions import ConnectionError
 
-from library.weaviate_schemas import WeaviateSchemas
+from library.models.weaviate_schemas import WeaviateSchemas
 from tests.integration.library.integration_test_base import IntegrationTestBase
 
 class TestSlackWeaviate(IntegrationTestBase):
